+++ tests/test_depth_interp.py
@@ -1,6 +1,5 @@
 # encoding: utf-8
-import common
-from common import outdir, TRAVIS
+from .common import outdir, TRAVIS
 import logging as ll
 import unittest as ut
